Move `to_xml` to `test_utils` so it can only be used for testing
